pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "streamlit>=1.48.1",
    "ta>=0.11.0",
    "tensorflow>=2.20.0",
    "websocket-client>=1.8.0",
]
